Changes for DAQKit 4.2

+++ emuDQM/dduBinExaminer/include/dduBinExaminer.hh
@@ -2,6 +2,7 @@
 #define OStream_hh
 #include <string>
 #include <stdio.h>
+#include <stdint.h>
 #include <iostream>
 
 class OStream : public std::ostream {
@@ -42,25 +43,25 @@ public:
 
 class dduBinExaminer {
 public:
-	const unsigned short nERRORS, nWARNINGS;
+	const uint16_t nERRORS, nWARNINGS;
 
 private:
 	std::vector<char*> sERROR,  sWARNING, sERROR_,  sWARNING_;
-	long               bERROR,  bWARNING;
+	int32_t               bERROR,  bWARNING;
 	bool               fERROR  [25];//[nERRORS];
 	bool               fWARNING[5]; //[nWARNINGS];
 
 	std::set<int>      fCHAMB_ERR[25]; // Set of chambers which contain particular error
 	std::set<int>      fCHAMB_WRN[5];  // Set of chambers which contain particular warning
-	std::map<int,long> bCHAMB_ERR;     // chamber <=> errors in bits
-	std::map<int,long> bCHAMB_WRN;     // chamber <=> errors in bits
+	std::map<int,int32_t> bCHAMB_ERR;     // chamber <=> errors in bits
+	std::map<int,int32_t> bCHAMB_WRN;     // chamber <=> errors in bits
 
 	OStream cout, cerr;
 
 	int currentChamber;       // ( (CrateNumber<<4) + DMBslot ) specifies chamber
 
-	const unsigned short *buf_2, *buf_1, *buf0, *buf1, *buf2;
-		  unsigned short tmpbuf[16];
+	const uint16_t *buf_2, *buf_1, *buf0, *buf1, *buf2;
+		  uint16_t tmpbuf[16];
 
 	bool fDDU_Header;
 	bool fDDU_Trailer;
@@ -77,49 +78,49 @@ private:
 	int  DAV_DMB;  // ...
 
 public:
-	long cntDDU_Headers;
-	long cntDDU_Trailers;
-	std::map<int,long> cntCHAMB_Headers;
-	std::map<int,long> cntCHAMB_Trailers;
+	int32_t cntDDU_Headers;
+	int32_t cntDDU_Trailers;
+	std::map<int,int32_t> cntCHAMB_Headers;
+	std::map<int,int32_t> cntCHAMB_Trailers;
 
 private:
-	long DDU_WordsSinceLastHeader;
-	long DDU_WordCount;
-	long DDU_WordMismatch_Occurrences;
-	long DDU_WordsSinceLastTrailer;
+	int32_t DDU_WordsSinceLastHeader;
+	int32_t DDU_WordCount;
+	int32_t DDU_WordMismatch_Occurrences;
+	int32_t DDU_WordsSinceLastTrailer;
 
-	long ALCT_WordsSinceLastHeader;
-	long ALCT_WordCount;
-	long ALCT_WordsExpected;
+	int32_t ALCT_WordsSinceLastHeader;
+	int32_t ALCT_WordCount;
+	int32_t ALCT_WordsExpected;
 
-	long TMB_WordsSinceLastHeader;
-	long TMB_WordCount;
-	long TMB_WordsExpected;
-	long TMB_Tbins;
-	long TMB_WordsExpectedCorrection;
+	int32_t TMB_WordsSinceLastHeader;
+	int32_t TMB_WordCount;
+	int32_t TMB_WordsExpected;
+	int32_t TMB_Tbins;
+	int32_t TMB_WordsExpectedCorrection;
 
-	long CFEB_SampleWordCount;
-	long CFEB_SampleCount;
-	long CFEB_BSampleCount;
+	int32_t CFEB_SampleWordCount;
+	int32_t CFEB_SampleCount;
+	int32_t CFEB_BSampleCount;
 
 	bool checkCrcALCT;
-	unsigned long ALCT_CRC;
+	uint32_t ALCT_CRC;
 	bool checkCrcTMB;
-	unsigned long TMB_CRC;
+	uint32_t TMB_CRC;
 	bool checkCrcCFEB;
-	unsigned long CFEB_CRC;
+	uint32_t CFEB_CRC;
 	bool checkCrcDDU;
-	unsigned long DDU_CRC;
+	uint32_t DDU_CRC;
 	//int headerDAV_Active; // Obsolete since 16.09.05
 
 public:
 	OStream& output1(void){ return cout; }
 	OStream& output2(void){ return cerr; }
 
-	long check(const unsigned short* &buffer, long length);
+	int32_t check(const unsigned short* &buffer, int32_t length);
 
-	long errors  (void) const { return bERROR;   }
-	long warnings(void) const { return bWARNING; }
+	int32_t errors  (void) const { return bERROR;   }
+	int32_t warnings(void) const { return bWARNING; }
 
 	const char* errName(int num) const { if(num>=0&&num<nERRORS)   return sERROR[num];   else return ""; }
 	const char* wrnName(int num) const { if(num>=0&&num<nWARNINGS) return sWARNING[num]; else return ""; }
@@ -133,18 +134,18 @@ public:
 	std::set<int> chambersWithError  (int num) const { if(num>=0&&num<nERRORS)   return fCHAMB_ERR[num]; else return std::set<int>(); }
 	std::set<int> chambersWithWarning(int num) const { if(num>=0&&num<nWARNINGS) return fCHAMB_WRN[num]; else return std::set<int>(); }
 
-	long errorsForChamber(int chamber) const {
-		std::map<int,long>::const_iterator item = bCHAMB_ERR.find(chamber);
+	int32_t errorsForChamber(int chamber) const {
+		std::map<int,int32_t>::const_iterator item = bCHAMB_ERR.find(chamber);
 		if( item != bCHAMB_ERR.end() ) return item->second; else return 0;
 	}
 
-	long warningsForChamber(int chamber) const {
-		std::map<int,long>::const_iterator item = bCHAMB_WRN.find(chamber);
+	int32_t warningsForChamber(int chamber) const {
+		std::map<int,int32_t>::const_iterator item = bCHAMB_WRN.find(chamber);
 		if( item != bCHAMB_WRN.end() ) return item->second; else return 0;
 	}
 
-	std::map<int,long> errorsDetailed  (void) const { return bCHAMB_ERR; }
-	std::map<int,long> warningsDetailed(void) const { return bCHAMB_WRN; }
+	std::map<int,int32_t> errorsDetailed  (void) const { return bCHAMB_ERR; }
+	std::map<int,int32_t> warningsDetailed(void) const { return bCHAMB_WRN; }
 
 	void crcALCT(bool enable);
 	void crcTMB (bool enable);
